Fix use of bitmap fonts on some Linux systems
In 30bd59a1dec78 we added a fallback from font config
to XLFD in the case where font config could not supply
us with the correct pixel size for a given font.
Primarily, this was for when you would try to print
a bitmap font on a high resolution printer. This caused
a problem for applications that set a point size on
a bitmap font, since the calculated pixel size would
not be available. For UIs you will usually prefer to
select the font with the correct family but nearest
pixel size here, which is what font config does. We
would however choose the fallback and if XLFD failed
to load the font in question, you would get a
substitution. Since getting the correct font family
seems more important than getting the correct pixel
size, we disable the fallback in the case where XLFD
does not find the correct font. This will fix the bug
where UI fonts were frequently created with the wrong
family on e.g. Fedora, and it might cause some trouble
with printing bitmap fonts, however, it was deemed
that bitmap fonts are not ideal fonts to use for
printing. In cases where XLFD can load the bitmap
fonts, the fallback will still be used.

Task-number: QTBUG-4428
Reviewed-by: Trond