Make qmlpixmapcache thread shutdown cleaner Change to Thiago's method:
QThread::quit() is *not* thread-safe, so we mustn't call it from
outside the QThread that the QThread holds. Unlike other QObjects,
this method and some others must not be called from the thread they
have affinity to.

So add a hack to call it in the auxiliary thread. Do that by
connecting a dummy QObject's destroyed() signal to the quit() slot in
a DirectConnection, then move the object to the thread. When we call
deleteLater(), this will cause the thread's event loop to delete the
object, which in turn calls quit() in the actual thread.

Reviewed-by: Thiago Macieira