Added additional logging to Phonon MMF backend
In addition to dumping the widget tree which contains the VideoOutput
widget, the backend now also dumps the widget tree from the application's
active window.  This is to allow debugging of problems which stem from
the fact that the VideoOutput widget subtree root is not the
application's main window.

In addition to dumping the Qt widget tree, the backend now also requests
the window server to dump the entire window tree.

Together, these changes allow easier debugging of video visibility-
related problems.

Reviewed-by: trustme