Fix QDate::isLeapYear() for years < 1
The current formula for isLeapYear() assumes that years -4, -8, etc are
leap years in the Julian Calendar. This is not the case, leap years in
fact fall in -1, -5, -9, etc as there is no year 0 in the Julian
Calendar (count back 4 years from 4AD to confirm). The
julianDayFromDate() and getDateFromJulianDay() functions correctly
calculate this, but isValid() uses isLeapYear() and so causes setDate()
to incorrectly reject valid dates.

Sample test code to round-trip a jd to ymd and back to jd:

  QDate testDate;
  QDate loopDate = QDate::fromJulianDay(1);
  while ( loopDate.toJulianDay() <= 5373484 ) { // <= 9999-12-31
    testDate.setDate( loopDate.year(),
                      loopDate.month(),
                      loopDate.day() );
    if ( !testDate.isValid() || loopDate != testDate ) {
      qDebug() << "Round Trip failed : " <<
                  loopDate.toJulianDay() << " = " <<
                  loopDate.year() << loopDate.month() <<
                  loopDate.day() << " = " <<
                  testDate.toJulianDay();
    }
    loopDate.addDays(1);
  }

Before the fix, 29 Feb 1 BC (year = -1) and all other BC leap days
returned by year() month() day() would result in setDate() failing.
After the fix the round trip works fine.

Merge-request: 2282
Reviewed-by: Olivier Goffart <ogoffart@trolltech.com>