QS60Style: Custom QPushButton: Heigth is calculated wrongly
If button contains a layout with widgets instead of having direct label
and icon, QS60Style calculates the size of the content incorrectly.

This is regression for a fix to QTBUG-7586. Now style ignores orginal
size of pushbutton and tries to calculate correct height by itself.

Issue is fixed by selecting teh bigger of orginal height and calculated
height.

Task-number: QT-3197
Reviewed-by: Alessandro Portale