[plugins] Don't conclude that a .debug file is a plugin.
On Unix systems (that are not Mac), the Qt buildsystem splits the code
from the debug symbols using objcopy. The resulting .debug file
contains a valid ELF header, with a copy of the main library's section
headers. But the contents of those sections are missing.

When we try to dlopen(3) those libraries, libdl crashes.

So we shouldn't try to open those files.