Fix crash for multiscript text in QTextEngine::boundingBox()
The logClusters pointer is already adjusted for the item, so it should
be indexed relative to the item, hence from 0 to itemLength-1. Adding
pos to the index would give us a random value for all script items > 0.
Thus this would sometimes crash for text that spanned several scripts,
since we would sometimes get a value back which was outside the bounds
of the glyph array.

Task-number: QTBUG-9374
Reviewed-by: Samuel