Fix painting artifacts on text
The height of the row is not only decided by the maximum height of the
set of glyphs currently being added to the glyph cache, but also the
height of the glyphs previously added to the current line in the glyph
cache. To minimize the space used, we now calculate the maximum height
for glyphs per line, and use this as the vertical advancement when
positioning the next line in the cache. The change also removes a
redundant test ((c_x + c.w > c.h) is caught earlier on and will
always be false at this point.)

Task-number: QTBUG-9176
Reviewed-by: Trond