eliminate special splitting of INCLUDEPATH and DEPENDPATH
it wouldn't work as expected anyway, as the splitting rule (using the
semicolon in addition to whitespace) applied only to the string literal
from the pro file, but not any expanded [environment] variables, etc.
(i.e., where it might make any sense). so just drop it, as it would
considerably complicate later optimizations. it wasn't documented
anyway ...

Reviewed-by: joerg