Carbon: problems closing down native applications
When Qt is used as a plugin, Qt always eats the quit event in the
application event handler. This patch makes sure we let the event
through in that case. And also that we don't catch the following
quit apple event. The whole buisiness of quitting down the native
application is not the responsibility of Qt when used as a plugin.

Reviewed-by: prasanth