Revert "Add QListModelInterface::modelReset() signal and emit this in"
This reverts commit 973cfce37fcdd1ce330f237eaa76930db55a73f6.

Need more consideration before adding modelReset(). For example if
item insertion/removal is supposed to trigger animations through
added/removed signals, they won't be triggered if only the modelReset()
signal is emitted. Also if we add modelReset(), it should also be
implemented for ListModel to make sure it is implemented by all subclasses
of QListModelInterface and to test the impact of this on the view classes.