Add SmoothedFollow element to qml
The SmoothedFollow is the same as the old EaseFollow, so it's not an
animation, but its main use case is to be used as a property value
source to automatically follow the 'to' property, as in the example below.

Rectangle {
    color: "green"
    width: 60; height: 60;
    SmoothedFollow on x { to: rect1.x - 5; velocity: 200 }
    SmoothedFollow on y { to: rect1.y - 5; velocity: 200 }
}

This element shares the internal implementation with SmoothedAnimation,
both providing the same easing function, but with SmoothedFollow it's
easier to set a start value to animate intially and then start to follow,
while SmoothedAnimation is still convenient for using inside Behaviors
and Transitions.

Reviewed-by: Michael Brasser