Fixed focus and window activation events on Symbian when opening menu.
As described in QTBUG-8698, Qt for Symbian has been generating
incorrect focus and window activation events. This has happened since
launching menu from QSoftkeyManager with TryDisplayMenuBarL, invokes
eventually QSymbianControl::FocusChanged. But when the FocusChanged is
called menu being launched is not yet set to visible, meaning that
IsDisplayingMenuOrDialog returns false.

Because there is no way in platform to detect that menu is being
launhced, the fix is to add a new flag QS60Data, which can be used to
detect if FocusChanged event is received due to the fact that menu
is being constructed/launched.

Task-number: QTBUG-8698
    * Fixes issues 2, 3 and 4
Reviewed-by: Sami Merila