Cleanup handling of errors in bindings and scripts
QML used to silently ignore a log of errors - such as a failed assignment
to a QObject property.  These errors are now all reported as exceptions
in JavaScript.

Other questionable activities, like assigning a JavaScript array to a
"property var" property which appeared to work, thanks to QtScript's
transparent conversion of arrays to a QVariantList, are now blocked
entirely.

QTBUG-9152 QTBUG-9382 QTBUG-9341 QTBUG-6886