Fix a crash when unloading libQtCore
QThread will register a cleanup function with pthreads to be run when
ever a thread exits. However, if QtCore is unloaded, this function would
stay registered and the application would crash when a thread exits
(since the function pointer registered with pthreads is no longer
valid).

Add a destructor function to qthread_unix.cpp which destroys the
pthread_key_t and unregisters the cleanup function.

Task-number: QTBUG-9436
Reviewed-by: Marius Storm-Olsen