Tune plugin import mechanism
In shadow build environments, we cannot enforce that shared
library objects for plugins are located in the same directory
as the qmldir file. This makes it hard for Creator to support
mixed projects (qml/c++).

In order to gain more flexibility, the patch introduces a
pluginPathList to QDeclarativeEngine, which completes the existing
importsPathList.

The pluginPathList defaults to ["."], which indicates the directory
where the qmldir file is located in.

The qml viewer tool gains a command line option -P to add to the
pluginPathList.

For consistency, the -L option ("Library") has been renamed to
-I ("Import"). QDeclarativeEngine::importExtension() has been
renamed to QDeclarativeEngine::importPlugin().

The documentation has been adjusted accordingly.

Done with erikv.

Reviewed-by: erikv