Some cleanup in widgets module. Deletion of GraphicsObjectContainer.
This commit deletes the uneeded classes/declarations since we have now
an inline support for QGraphicsObject based classes. QGraphicsLayout
bindings works the same way because the layout property has been added
to QGraphicsWidget. The only feature that is missing for now (because
of the deletion of GraphicsObjectContainer) is the anchoring support.
We will probably take a look on how to support this feature properly
with no wrapper.

Task-number:QTBUG-9456
Reviewed-by:Michael Brasser