Simplify how the QGLEngineSharedShaders object is cleaned up.
There are two things that needs to be done when cleaning up those
objects: 1) if the context is destroyed before the specific thread
exits, the shaders that is allocated within that context needs
to be cleanup before the context is destroyed. 2) if the thread
exits before the context is destroyed, the thread needs to clean
up the shaders and then delete the QGLEngineSharedShaders object.

We can't use a QGLContextResouce for that because a) we don't know
if our context ptrs are valid, b) QGLContextResource doesn't remove
themselves from the QGLContextGroups internal list when destroyed,
so we might get a free() call *after* the resource had been destroyed.
Use the contextDestroyed() signal from QGLSignalProxy() instead, which
is cleaner and actually works..