Fix crash with QTextEdit::textChanged() when deleting a character
QTextEdit::textChanged() will be emitted from a function called from
within QTextCursorPrivate. If the code connected to textChanged()
makes a local copy of the current text cursor and then causes it
to detach, we will crash when returning to QTextCursorPrivate and
trying to access the now-deleted data. To avoid this, we make a local
reference to the current text cursor that gives us a guarantee that
it will be valid throughout the delete-call.

Task-number: QTBUG-9599
Reviewed-by: Gunnar