Improved support for OPTION and LINKEROPTION statements in MMP files
VERSION_FLAGS.<keyword> can now be used for all compilers and not
just armcc, and version flagging can now be used with QMAKE_LFLAGS
as well as QMAKE_CXXFLAGS.

Also, MMP_OPTION_KEYWORDS variable is used to define supported keywords
for OPTION and LINKEROPTION statements, which are defined via
QMAKE_CXXFLAGS.<keyword> and QMAKE_LFLAGS.<keyword> variables.
This improves flexibility in the future if new keywords need to be
supported.

Task-number: QTBUG-8685
Reviewed-by: Janne Koskinen