Make debugging floating point mistakes much easier.
Its a common mistake that a floating point error, like divide by zero,
is not detected and just propagated till it breaks in curious ways elsewhere.
Like QBezier hitting an infinite loop due to operating on NaN floating points.
Adding detection of this in a debug build of Qt is sure to help a lot of these
issues be detected much faster and so Qt helps people write code even faster.

Reviewed-By: Gunnar Sletta