Fix a race where QThread::exit() is "lost" when called after start()
QThread::exit() tries to stop all running event loops, but does nothing
if the eventloop has not started yet. This is often the case for short-
exit method on an object that has affinity to the thread. This ensures
that the exit is called from the running eventloop, meaning the exit()
will never be lost.

Task-number: QTBUG-1184
Reviewed-by: Morten Sørvig