CSS: fixes border only affecting the first widget.
Rules like  "Foo { border: 2px solid; }" does not specify the border color.
When the border color is not specified, it is assumed to be black.

When reading the brush value from the cache, we should take that
into account.

Note that this logic cannot be moved into brushFromData() as it is different
for the background. (when no color is specified, it is assumed to be transparent)

Reviewed-by: jbache
Task-number: QTBUG-9674 (part one)