Implement some changes to the AIX xlC mkspec suggested by IBM.
Set -bmaxdata to 0x8000000 for normal 32-bit Qt programs, to allow
them to access more memory. Increase that limit to the maximum allowed
when linking to QtWebKit (even though we don't support WebKit with xlC,
there are patches to do that).

For 64-bit, simply add the "big TOC" flag, which enables accessing
more symbols that cannot be reached by a 16-bit addressing. Only
QtWebKit strictly needs it, but IBM suggests as a good flag for
everyone.

Reviewed-by: Thomas Zander