Skip test for cocoa: setToolTip
We need to skip the QWidget::setToolTip test for now since it fails
on Cocoa: The reason is that the test uses QCursor::setPos that ends
up faking a mouse move on mac (which triggers tooltips). And when
the mouse moves, a mouse enter event is delivered to the widget
through cocoa, which will also fake a mouse move event. Until we
know what the best solution is, we chose to skip this test

Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@nokia.com>