Prevent accelerators from firing on X11 and QWS during a drag
Drag-and-drop has always been modal, and already blocks key events. It
should block all application accelerators as well, to avoid unexpected
things like being unable to cancel the drag-and-drop (as mentioned in
the task), or by being able to unexpectedly recurse the drag-and-drop
event loop (by, for example, opening a file dialog).

Task-number: QTBUG-9786
Reviewed-by: TrustMe