Bring support of anchors in QML for QGraphicsWidget derived classes.
This commit add a extension object to bring the anchors property for
a QGraphicsWidget. The actual implement uses a separate object for storing
the anchor pointer. In the future it would be nice if the extension object
is the anchor itself. Also there are two connects extra for QGraphicsWidget
one can perhaps disappear with a later commit. Only baseline case is not
supported because QGraphicsWidget don't have any concept of baseline.

Reviewed-by:akennedy