tst_mediaobject: ensure MediaObject is in StoppedState before each step
Many of the steps in the tst_mediaobject suite (a) check that the
MediaObject is in StoppedState at the start of the step and (b) call
stopPlayback() at the end.  If, however, a QTest check fails during
the test, stopPlayback may not be called.  This patch adds a call to
MediaObject::stop() in the suite's init() function.

This is a symptom of a wider problem with this test suite, namely that
it re-uses a single instance of Phonon::MediaObject for all steps.  Given
the highly stateful nature of MediaObject, this can lead to test steps
failing due to some state which was erroneously carried forward from an
earlier step.  While this test suite design may more faithfully
represent real-world usage of Phonon, it makes tracking down the root
causes of test failures needlessly difficult.

Reviewed-by: Frans Englich