QtScript: Fix call stack issue with qsTr() when JIT is enabled
When the JIT is enabled, as well as JIT_OPTIMIZE_NATIVE_CALL defined,
NativeFunctionWrapper is a typedef for JSFunction. This has the
consequence that the native call frame will not be fully initialized
by JIT-generated code (because it shouldn't have to); in particular,
ExecState::codeBlock() is not set up.
qsTr() relies on codeBlock() to return a sensible value, though, so
it breaks this contract.

By making qsTr a PrototypeFunction, the JIT will detect that the
function call needs more elaborate setup, i.e. initialize codeBlock()
as well.

Reviewed-by: Olivier Goffart