Phonon MMF: fixed crash during opening of video clip
A crash was observed during opening a video clip, and was traced
to dereferencing a null m_player pointer in
DsaVideoPlayer::handleParametersChanged(), which is called during
construction, but before createPlayer() has been called.

This was reproducible using the following sequence:
    1. Launch qmediaplayer
	2. Play an audio clip
	3. Open a video clip

However, the following sequence worked as expected:
    1. Launch qmediaplayer
	2. Play a video clip
	3. Play an audio clip
	4. Play a video clip
	...

It is not clear which commit introduced this defect.

Reviewed-by: Frans Englich