Slider graphics does not look correct in N95
The upscaling of theme graphics fails, if the original rect is taller
than upscaled graphics rect height. This makes the start and end parts
of the graphics appear as shorter then they should be.

Fixed by forcing the destrination rect height to be used after
upscaling the graphics.

Task-number: QTBUG-9854
Reviewed-by: Janne Anttila