Fix tst_QEventLoop::exec() regression introduced by commit 816523117bc00cfeb17e347f7fe5f11278a5e871
The quitNow flag needs to be reset at the beginning of exec() to allow
exec() to recurse and be run multiple times.

This changes reverts commit 816523117bc00cfeb17e347f7fe5f11278a5e871 an
introduces the QThreadPrivate::exited flag to fix the race between
start() and exit(), and QThreadPrivate::returnCode to make sure exec()
returns the code passed to exit().

Task-number: QTBUG-1184
Reviewed-by: olivier