Re-apply change 81837e43e3f966c1755e90eb65df6e3bad506ed7 by Iain
Symbol visibility fixes for RVCT4 on Symbian
RVCT 4 is far more strict with regards to symbol visiblity that RVCT
2.2, and will hide symbols unless all references have default visibility
in the object files.  Update the various places in Qt code where the
symbol visibility was set incorrectly for DLL-based platforms (those that
use __declspec(dllimport) and (dllexport).

Note: Other Qt modules and QtScript are fixed in different commits.

Task-number: QTBUG-9903
Reviewed-by: Jason Barron
             Janne Koskinen