QScript: use JSC::NativeFunctionWrapper instead of JSC::PrototypeFunction when possible
JSC::NativeFunctionWrapper is a typedef to either JSC::PrototypeFunction or
JSC::JSFunction depending if we are running JIT or not.

When using JIT, JSC::JSFunction is faster, as it allow JIT to do the native
call dirrectly.  The difference is that in that case, the JS stack is not
fully set up so we have to be carefull.

Unfortunately, it is not possible to make FunctionWrapper inherit from
JSC::NativeFunctionWrapper, because JSFunction is slightly bigger, and we
cannot fit in a Cell

Reviewed-by: Kent Hansen