Fixed parsing of inner labelled statements.
The JS grammar is ambigious and the following statement can be parsed
as an object-literal followed by an inserted semicolon or as two
labelled statements.

outer: {
  inner: {}
}

In the old days we used to resolve the conflict by reducing the
statement to an expression statement but this was wrong so
now we prefer the labelled statement. As nice side effect, we
pass two more tests in tests/auto/declarative/parserstress.

Task-number: QTBUG-8108