QS60Style: Qt does not draw transparency correctly
This is due to default mask depth in QS60Style. It is currently set as
default EGray2 and then inquiried from S60SkinServer. For 3.1 devices,
skinserver does not explicitly set the mask depth (even though default
is 8bit mask), so no value is returned and default value is used.
This leads to a situation where style uses 2bit masks for 9 or 3 part
frame graphics.

Corrected by changing the default value to 8bit mask (EGray256).

Task-number: QTBUG-9927
Reviewed-by: Janne Anttila