Mac: Fix off-by-one in vertical position for elided and non-elided text
In the code path that draws the elided text, we would truncate the
position before passing it to the painter. With a font engine that
supports fractional values (mac), this would potentially give us the
wrong position compared to the code path that draws the complete text,
which essentially rounds off the number. The result was that when you
resized the width an item view to make its items elide the text, then
they would potentially shift up or down by one pixel.

Task-number: QTBUG-9879
Reviewed-by: Gunnar