Don't use setGeometry() in setWindowState() on Symbian.
setGeometry() has the side effect that it internally sets both the
WA_Moved and WA_Resized attributes and these attributes should only be
used when the application has explicitly set a size or position. When
a widget is resized due to being maximized or fullscreened then we
should resize the native window handle instead of calling setGeometry.

Reviewed-by: Janne Anttila