Delay creation of S60 screen furniture.
Previously the S60 screen furniture (Status Pane and CBA) was being
constructed when QApplication was constructed. This meant that if you
never created a widget, you would still get those app panes which is
not what we want.

This patch avoids constructing these panes in QApplication, and manually
creates them later when the first window is shown in a state where they
would be needed.

Reviewed-by: Janne Anttila
Reviewed-by: mread