Use QDesktopWidget as a status pane observer on Symbian.
Now that the S60 screen furniture construction is delayed, we need to
manually set the status pane observer to get notifications about
changes in the size of the status pane. Using the AppUi would have been
the obvious choice here, but sadly CAknAppUi (base class for
QS60MainAppUi) uses private inheritance when implementing the
MEikStatusPaneObserver interface.

To work around this problem we make QSymbianControl implement the
interface and use the instance of this class that is associated with
QDesktopWidget to recevive the notifications and then call the function
in the AppUi to maintain the same behavior.

Reviewed-by: Janne Anttila
Reviewed-by: mread