Tweak the 'normalGeometry' of the widget before setting it.
If a widget is initially shown as maximized or fullscreen then the
normal geometry is calculated before the status pane has been created.
When returning to the "Normal" window state the geometry needs to be
tweaked to compensate the fact that there is a status present now. This
should only happen if the application has not been given an explicit
position.

Reviewed-by: Janne Anttila