Fix crash on startup on Symbian OS
The changes to QThread introduced by commit
9aa4538b219ed759a47e8d1f93c2797bf07b5e2f mean that the QThread constructor
calls into the event dispatcher.
The Symbian event dispatcher owns a QThread, so it crashed when the code
re-entered the partially constructed event dispatcher and used an
uninitialised pointer.

This change delays construction of the QThread until the point of use,
so that the event dispatcher is fully constructed.

Task-number: QTBUG-10029
Reviewed-by: Jason Barron
(cherry picked from commit 2b55d52669beb72396f94e449fdf172735349b3b)