No repaint when resizing graphics item with an effect.
Problem was that the item's cache was not invalidated from
prepareGeometryChange(). We did invalidate the cache for all the
ancestors, but not for the item itself. To avoid looping through the
whole parent chain twice, we invalidate the item's cache before
walking the parent chain. It would be nice to centralize the cache
invalidating mechanism, but for now it's OK to solve it like this.

Auto test included.

Task-number: QTBUG-9551
Reviewed-by: samuel