Fix window transparency on Symbian.
In the create_sys() function where native handles are created, we were
only looking at the "isOpaque" flag to determine if a window should be
translucent or not. This is not the correct thing to do because
transparency should only be granted if the application explicitly sets
the Qt::WA_TranslucentBackground flag.

The fix is to defer the transparency decision to s60UpdateIsOpaque()
which does the right thing.

Reviewed-by: Jani Hautakangas