Default (Parentless) QMenuBar actions do not work
This bug is a bit nasty. The problem is that Cocoa is trying
to be smart behind our back, and enable/disable menu items
depending on the menu item targets. But in Qt, we have a
different policy that Cocoa on when a menu bar should be
disabled or not. E.g. we allow a modal dialog to access the
application menu bar if it is the only window on screen.

This patch will work around cocoa being smart by setting the
menu item targets dynamically, depending on the modal state
of Qt. Setting it to nil will make the items enabled when there
is a modal dialog on screen, and setting it to the menu loader
will enable it when there is _no_ window on screen at all.

Task-number: QTBUG-9209
Reviewed-by: prasanth