Optimize ~QObject
By avoiding to lock mutextes when not needed.

According to experiment, many object are destroyed without any connection
attached to them. (80% with qml, 50% with creator)

d->currentSender does not need to be protected by the mutex since
we are in the destructor.

We do not need Q_TRY since the mutex is grenteed to exist if there
is connections.

We do not need to call QCoreApplication::removePostedEvents if
there is not events to process. And d->postedEvents can be checked
outside of the lock since it cannot be incremented anymore in the destructor

Reviewed-by: Roberto Raggi
Reviewed-by: Brad