QX11GL: Don't do glFinish in endPaint
It's better to defer the synchronisation to just before the point it
is actually needed. When used as a window surface, this is in flush
and scroll. If the QX11GLPixmapData is used as the defaut backend
for QPixmaps, it might need to be put back into endPaint. However,
the GL driver will hopefully make sure rendering to the pixmap is
complete before binding it as a texture via texture-from-pixmap.

Also, it's probably better to use eglWaitClient rather than glFinish
for synchronisation as it is potentially slightly more optimal.

Reviewed-By: TrustMe