Make tst_qdialog::throwInExec work with symbian exception policy
On Symbian, exceptions are absorbed by the event loop. This behaviour
is inside the user library where we can't modify it.
Only std::exception or leaves are supported, other exceptions will cause
an assert fail in the symbian code. So the test is changed to throw a
std::exception derived class

Reviewed-by: mread