Inline the common case when there is no contention.
QMutex locking could be used in performence critical code (such
as QObject and co.) where the cost of a function call is
not neglectible. This would slow down the common case where
there is only one thread and no contention.

Since we need to keep the old symbol for binary compatibilty, introduce
new *Inline function.

Reviewed-by: Brad