When using Qt::BypassGraphicsProxyWidget with QMenu the application is not stuck anymore
When using this flag the child of the widget which is embedded is becoming
a top level QWidget. So a right click on a text edit trigger the context
menu to be top level. When creating this top level context menu we are
grabbing the mouse but we were never releasing it when the menu was hidden.
The patch check if the widget uses the Qt::BypassGraphicsProxyWidget and
ungrab the mouse. The patch also fix a bug when positioning the QMenu, it was
for the same reason.

Task-number:QTBUG-7254
Reviewed-by:brad