Fix QDeclarativeTextInput and QGraphicsView in regards of input methods hints.
When the echoMode changes for QDeclarativeTextInput we need to update
the inputmethod hints.
Also in QGraphicsView we need to update the input context installed on the
view to make the former is aware of the change.
For the input context framework the only way to deal with the current widget
is focusWidget(). Unfortunately the widget that has the focus is the
QGraphicsView so the input context don't know the real object QGraphicsView
is actually focusing. We must keep in sync the inputMethodHints of QGV
with the object QGraphicsView is focusing so the input context just
called focusWidget()->inputMethodhints() to update itself.

Task-number:QTBUG-9922
Reviewed-by:janarve
Reviewed-by:bnilsen
Reviewed-by:Michael Brasser