Fix update issue when an item has an effect and child.
When a child marks its parent as dirty (and the parent has an effect) and they are
not in a scene, they must not set fullUpdatePending to 1 because if the scene
is in fullUpdate (not yet painted) then the item will not be processed by
processDirtyItemRecursive so the fullUpdatePending flag will never be
reset by resetDirtyItem. Any calls to update() will be then discarded
because it will think that a full update is pending for the item. A full
update will be done anyway by the scene and if not then markDirty will do
its job when the parent will be added to the scene.

Task-number:QTBUG-10037
Reviewed-by:bnilsen