Fix regression: auto completion text cursor problem in Q3FileDialog
This was another regression in QLineEdit introduced with the move to QLineControl

setCursorPosition clear the selection, and setSelection move the cursor position.
So the case where we want the position at the beginning of the selection need
to be handled separately (by doing the selection reverse)

This does not cover the case where the position is in the middle, or outside
the selection, but it is not possible to do.

Task-number: QTBUG-10019
Reviewed-by: jbache