Implement heightForWidth support for QTabWidget and QStackedWidget.
The problem was simply that QTabWidget does not have a layout (it
does manual layouting, and since that the heightForWidth implementation
it inherits from QWidget require that the widget has a layout it will
not work and it always returned -1 for QTabWidget). The solution is then
to reimplement heightForWidth().

Unfortunately the patch has several workarounds for BIC:
1. QWidget::hasHeightForWidth() should really have been virtual. Instead
   we add the virtual QWidgetPrivate::hasHeightForWidth().
   Since this is a workaround/bugfix for QTabWidget we decided to keep
   QWidget::hasHeightForWidth() internal for now.

2. We cannot reimplement a virtual function in QStackedLayout.
   We therefore fix QStackedWidget by subclassing QStackedLayout and
   reimplement the virtual functions in the subclass.

This is not an ideal fix, but improves QTabWidget and QStackedWidget
wrt height for width.

Task-number: QTBUG-7792
Reviewed-by: Paul
Reviewed-by: Olivier