Stub implementations for EGL for symbian
This is done to keep binary compatibility between Qt built with openvg
vs Qt built without openvg support.
The problem is that Symbian uses .def files to map between exported
symbol names and ordinals in the DLL export table. The alternative of
manually maintaining two versions of the QtGui def files proved to be
too error prone and time consuming.
Note that the EGL exports are defined in a private header, for use by
the openvg and opengl graphics system plugins.
These plugins should always be compiled against Qt configured with
support for the graphics system, as the headers contain default parameters
which are inlined into the plugin binary.

Task-number: QTBUG-7870
Reviewed-by: Tom Cooksey